refactor(widget): #24 Channel screen indentation
